Extract polling helpers into src/client/client-helpers.ts (warren-3f09)
Move PollUntilTerminalInput, pollUntilTerminal, sleepWithSignal and the
DEFAULT_POLL_INTERVAL_MS / DEFAULT_POLL_TIMEOUT_MS constants out of
client.ts into a new sibling client-helpers.ts. Re-export the two
constants from client.ts to preserve the public surface consumed by
src/client/index.ts. No behavior change.
